% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\docType{data}
\name{asr}
\alias{asr}
\title{\code{ResultSet} for testing and illustration purpouses}
\format{An object of class \code{ResultSet} of length 15.}
\usage{
data("asr")
}
\value{
A \code{ResultSet} object.
}
\description{
\code{ResultSet} created using \code{\link{association}} method, testing
 proteome association to exposome (\code{"mds"}), adjusted by sex and age.
}
\examples{
data("asr", package = "omicRexposome")
asr
}
\keyword{datasets}
