% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocES.R
\name{assocES}
\alias{assocES}
\title{Function to perform a Feature-Wide Association Study}
\usage{
assocES(x, y, formula, select, set = "exposures", sva = FALSE,
  vfilter = NULL, eBayes = TRUE, verbose = FALSE, warnings = TRUE, ...)
}
\arguments{
\item{x}{An \code{\link{ExposomeSet}} object or an
\code{\link{ExposomeClust}} object.}

\item{y}{An \code{\link{ExpressionSet}} object.}

\item{formula}{Formula indicating the covariates to be used in the
analysis.}

\item{select}{(optional) Character vector of exposures or fenotypes (see
argument \code{set}). If given, only this exposusures or phenotypes will be
tested, otherwise all the exposures or phenotypes in \code{x}  will be
tested.}

\item{set}{(default \code{"exposures"}) Set to be used for the association
analysis.}

\item{sva}{(default \code{FALSE}) If set to true, Surrogate Variable
Analysis will be done on transcriptome level in each test and the model
will be corrected by those surrogate variables.}

\item{vfilter}{(defaul \code{NULL}) Number of probes from transcriptome
to perform the Surrogate Variable Analysis. If \code{NULL}, all the probes
are used.}

\item{eBayes}{(default \code{TRUE}) If \code{TRUE} applies \code{eBayes}
after \code{lmFit}.}

\item{verbose}{(default \code{FALSE}) If set to \code{TRUE}, a series of
messages descriving the process are shown.}

\item{warnings}{(default \code{TRUE}) If set to \code{TRUE}, a series of
warnings are shown when required user atention.}

\item{...}{Arguments passed to \code{\link{lmFit}}.}
}
\value{
An object of class \code{\link{ResultSet}}.
}
\description{
This function allows to perform a Feature-Wide Association Study
by using an \code{ExposomeSet} and an \code{ExpressionSet}. The function
allows to perform the test to any continuous feature (aka. transcriptome,
methylome, proteome, etc.) It allows to perform an adjustment using Surrogate
Variable Analysis (from R package \code{sva}).
}
\examples{
library(rexposome)
data(gexp_r)
data(exp_r)
rst <- assocES(exp_r, gexp_r, formula=~sex+age)
rst
}
