% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{association}
\alias{association}
\title{# -----------------------------------------------------------------------------
Method to perform an association study between transcriptome and exposom}
\usage{
association(object, formula, expset, omicset, set = "exposures",
  method = "ls", ..., sva = FALSE, ebayes = TRUE, verbose = FALSE,
  warnings = TRUE)
}
\description{
This function allows to perform an association study between gene
expression from microarray and the exposome. An \link{ExpresionSet} is
the object storing the gene expresion and an \link{ExposomeSet} the one
storing the exposome. Both of them needs to be encapsulated in a
\link{MultiDataSet}. The association study is perform through standard
\link{limma} pipeline. The function allows to perform multiple tests using
the argument \code{exposures}.
}
