% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\docType{data}
\name{crs}
\alias{crs}
\title{\code{ResultSet} for testing and illustration purpouses}
\format{An object of class \code{ResultSet} of length 1.}
\usage{
data("crs")
}
\value{
A \code{ResultSet} object.
}
\description{
\code{ResultSet} created using \code{\link{crossomics}} method, selecting
 \code{"mcia"} method. Result from the integration of proteome and
 exposome data (\code{"mds"}).
}
\examples{
data("crs", package = "omicRexposome")
crs
}
\keyword{datasets}
