% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_r.R
\docType{data}
\name{exp_r}
\alias{exp_r}
\title{\code{ExpressionSet} for illustration purpouses.}
\format{An object of class \code{ExposomeSet} of dimension 6 x 109 x 9.}
\usage{
data("exp_r")
}
\value{
An \code{ExposomeSet} object.
}
\description{
\code{ExpressionSet} for illustration purpouses.
}
\examples{
data("exp_r")
exposureNames(exp_r)
}
\keyword{datasets}
