% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R, R/ResultSet-getIntegration.R
\docType{methods}
\name{getIntegration}
\alias{getIntegration}
\alias{getIntegration,ResultSet-method}
\alias{getIntegration}
\title{Method to extrat integration-feature result from a ResultSet}
\usage{
getIntegration(object, ...)

\S4method{getIntegration}{ResultSet}(object, ...)
}
\arguments{
\item{object}{An object of class \link{ResultSet} obtained from}

\item{...}{NOT USED}
}
\value{
A \code{data.frame}
}
\description{
Homologous methods from \code{MultiDataSet} (\code{getAssociation}) but
for \code{ResultsSet} created by \code{\link{crossomics}}. It Resturns a
\code{data.frame} with the result from \code{mcia} (\code{omicade4}) or
from \code{MultiCCA} (\code{PMA}).
}
\examples{
data("crs", package = "omicRexposome")
class(getIntegration(crs))
}
