% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp_r.R
\docType{data}
\name{gexp_r}
\alias{gexp_r}
\title{\code{ExpressionSet} for illustration purpouses.}
\format{An object of class \code{ExpressionSet} with 10000 rows and 75 columns.}
\usage{
data("gexp_r")
}
\value{
An \code{ExpressionSet} object.
}
\description{
\code{ExpressionSet} for illustration purpouses.
}
\examples{
data("gexp_r")
featureNames(gexp_r)
}
\keyword{datasets}
