% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\docType{data}
\name{mds}
\alias{mds}
\title{\code{MultiDataSet} for testing and illustration purpouses}
\format{An object of class \code{MultiDataSet} of length 2.}
\usage{
data("mds")
}
\value{
A \code{MultiDataSet} object.
}
\description{
\code{MultiDataSet} containing both proteome data-set and exposome
 data-set.
}
\examples{
data("mds", package = "omicRexposome")
mds
}
\keyword{datasets}
