% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methy_r.R
\docType{data}
\name{methy_r}
\alias{methy_r}
\title{\code{ExpressionSet} with methylation (beta) data for illustration purpouses.}
\format{An object of class \code{ExpressionSet} with 10000 rows and 200 columns.}
\usage{
data("methy_r")
}
\value{
An \code{ExpressionSet} object.
}
\description{
\code{ExpressionSet} with methylation (beta) data for illustration purpouses.
}
\examples{
data("methy_r")
featureNames(methy_r)
}
\keyword{datasets}
