% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicRexposome.R
\docType{package}
\name{omicRexposome}
\alias{omicRexposome}
\alias{omicRexposome-package}
\title{omicRexposome: Package for exposome and omic data associatin and integration}
\description{
omicRexposome: Package for exposome and omic data associatin and integration
}
\section{exposome-omic data association study}{

The packages offers the function \code{\link{assocES}} that allows to perform
an association study using transcriptome, methylome, etc. as dependent
variable and exposome data as independent variable. The function relies on
\code{limma} pipeline and generates an object of class \code{\link{ResultSet}},
that can be ploted using \code{\link{plotAssociation}}.
}

\section{exposome-omic data integration study}{

The packages offers the function \code{\link{crossomics}} that allows to perform
two types of integration study: Multi Canonical Correlation Analysis and
Multi Co-Inertia Analysis. The function allos to use any type and number of
datasets (aka. exposome transcriptome, methylome, etc.). The function generates an
object of class \code{\link{ResultSet}}, that can be ploted using
\code{\link{plotIntegration}}.
}

