% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{options}
\alias{options}
\title{Method to get the options sued to create the ResultSet}
\usage{
options(object)
}
\arguments{
\item{object}{A \code{\link{ResultSet}} object.}
}
\description{
Method that returns a list with the options used to create the
\code{ResultSet}.
}
