% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotAssociation}
\alias{plotAssociation}
\alias{plotAssociation,ResultSet-methods}
\title{Function to draw de result of an association study}
\usage{
plotAssociation(object, rid = 1, coef = 2, contrast = 1,
  type = c("manhattan", "qq"), tPV, tFC)
}
\arguments{
\item{object}{An object of class \link{ResultSet} obtained from assoc_*
functions.}

\item{rid}{(default \code{1}) Index or name of the test to be plotted. Not
used it the \code{ResultSet} comes from \link{assocSNP}.}

\item{coef}{(default \code{2}) Index of the coefficient to be extracted.}

\item{contrast}{(default \code{1}) When \code{code} corresponds to a
multicategorical variable, contasr selects the comparison.}

\item{type}{Can take \code{"manhattan"} or \code{"qq"}.}

\item{...}{Other arguments used in the different plots.}
}
\description{
This function draws two type of plots for the ResultSet from association
functions
}
\seealso{
\link{plotIntegration} for plotting integration results
}
