% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotLambda}
\alias{plotLambda}
\title{Plot lambda score for all results in a ResultSet}
\usage{
plotLambda(object, width = 0.75)
}
\arguments{
\item{object}{An object of class \link{ResultSet}}

\item{width}{(default \code{0.70}) width of the bar}
}
\value{
An lambda plot
}
\description{
This method draws a baplor with the lambda score of each result in the
given \link{ResultSet}.
}
\examples{
data(gexp_r)
data(exp_r)
rst <- assocES(exp_r, gexp_r, formula=~sex+age)
plotLambda(rst)
}
