% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotLambda}
\alias{plotLambda}
\title{Plot lambda score for all results in a ResultSet}
\usage{
plotLambda(object, trim = 0.5, width = 0.75)
}
\arguments{
\item{object}{An object of class \link{ResultSet}}

\item{trim}{(default \code{0.5}) percentage of right omited values for
\link{lambdaClayton}.}

\item{width}{(default \code{0.70}) width of the bar}
}
\description{
This method draws a baplor with the lambda score of each result in the
given \link{ResultSet}.
}
