% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_r.R
\docType{data}
\name{prot_r}
\alias{prot_r}
\title{\code{ExpressionSet} with proteome data for illustration purpouses.}
\format{An object of class \code{ExpressionSet} with 5 rows and 32 columns.}
\usage{
data("prot_r")
}
\value{
An \code{ExpressionSet} object.
}
\description{
\code{ExpressionSet} with proteome data for illustration purpouses.
}
\examples{
data("prot_r")
featureNames(prot_r)
}
\keyword{datasets}
