% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ResultSet-fData.R,
%   R/ResultSet-length.R, R/ResultSet-names.R, R/ResultSet-opt.R,
%   R/ResultSet-plotAssociation.R, R/ResultSet-plotHits.R,
%   R/ResultSet-plotIntegration.R, R/ResultSet-plotLambda.R, R/ResultSet-rid.R,
%   R/ResultSet-tableHist.R, R/ResultSet-tableLambda.R, R/ResultSet-topTable.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\alias{fData,ResultSet-method}
\alias{length,ResultSet-method}
\alias{names,ResultSet-method}
\alias{opt,ResultSet-method}
\alias{plotAssociation,ResultSet-method}
\alias{plotHits,ResultSet-method}
\alias{plotIntegration,ResultSet-method}
\alias{plotLambda,ResultSet-method}
\alias{rid,ResultSet-method}
\alias{tableHits,ResultSet-method}
\alias{tableLambda,ResultSet-method}
\alias{topTable,ResultSet-method}
\title{Class ResultSet}
\usage{
\S4method{fData}{ResultSet}(object)

\S4method{length}{ResultSet}(x)

\S4method{names}{ResultSet}(x)

\S4method{opt}{ResultSet}(object)

\S4method{plotAssociation}{ResultSet}(object, rid = 1, coef = 2,
  contrast = 1, type = c("manhattan", "qq"), tPV, tFC,
  show.effect = FALSE)

\S4method{plotHits}{ResultSet}(object, th = 0.05, width = 0.75)

\S4method{plotIntegration}{ResultSet}(object, cmpX = 1, cmpY = 2,
  lb.th = 0.2, legend.show = TRUE, colors, ...)

\S4method{plotLambda}{ResultSet}(object, width = 0.75)

\S4method{rid}{ResultSet}(object)

\S4method{tableHits}{ResultSet}(object, th = 0.05)

\S4method{tableLambda}{ResultSet}(object, trim = 0.5)

\S4method{topTable}{ResultSet}(object, rid, coef = 2, sort = TRUE)
}
\arguments{
\item{object}{A \code{ResultSet} object.}

\item{x}{A \code{ResultSet} object.}

\item{rid}{Name or index of the internal result to be used}

\item{coef}{Coefficient to be returne, usually 2}

\item{contrast}{If coefficient to be used was multicategorical, number
of the contrast to be returned.}

\item{type}{Type of plot to be drawn}

\item{tPV}{Threshold for P-Value}

\item{tFC}{Threshold for log FC of effect}

\item{th}{Threshold to considere a P-Value a hit}

\item{width}{Width of the bars}

\item{cmpX}{Component to be drawn on X-axis}

\item{cmpY}{Component to be drawn on Y-axis}

\item{lb.th}{Threshold to considere to place labels or not}

\item{legend.show}{Indicates if legend should be placed or not}

\item{colors}{Labeled vector for colors of the features}

\item{...}{Other arguments passed to internal methods}

\item{trim}{0.5}

\item{sort}{Indicates of the result should be ordered by P-Value}
}
\value{
An object of class \code{ResultSet}
}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \code{omicRexposome}. This include
the functions \code{\link{assocES}} and \code{\link{crossomics}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fData}: Returns \code{data.frame} with feature's data.

\item \code{length}: Returns the amoung of analyses stored in the
\code{ResultSet}.

\item \code{names}: Returns the names of the omics data used to create
the \code{ResultSet}.

\item \code{opt}: Returns a list with the options used to create the
\code{ResultSet}

\item \code{plotAssociation}: Allows to plot a series of plots (QQ plot, Manhattan
plot and Volcano plot) depending on the results stored in the
\code{ResultSet}.

\item \code{plotHits}: Draws a bar plot with the number of hits in each
result of the \code{ResultSet}.

\item \code{plotIntegration}: Draws a plot depending on the integration analysis
stored in the \code{Resultset}.

\item \code{plotLambda}: Draws a bar plot with the lambda value of each
analyses stored in the \code{ResultSet}.

\item \code{rid}: Returns the names (aka. exposures) of the analyses
stored in the \code{ResultSet}.

\item \code{tableHits}: Returns a table with the number of hits in each
result of the \code{ResultSet}.

\item \code{tableLambda}: Returns a table with the lambda value of each
analyses stored in the \code{ResultSet}.

\item \code{topTable}: Getter to obtain the raw \code{data.frame} from
association and integration analysis.
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{options}}{list of options used to create the \code{ResultSet}.}
}}

