% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{rid}
\alias{rid}
\title{Obtain the "rid"s from a ResultSet}
\usage{
rid(object)
}
\arguments{
\item{object}{An object of class \link{ResultSet}}
}
\value{
A character vector of \code{rid}s.
}
\description{
This method resunts as character with the \code{rid}
in a given \link{ResultSet}.
}
