% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{topTable}
\alias{topTable}
\title{Method to extrat feature result from a ResultSet}
\usage{
topTable(object, rid, coef = 2, sort = TRUE)
}
\arguments{
\item{object}{A \code{\link{ResultSet}} object.}

\item{rid}{The name or index of the result to be extracted.}

\item{coef}{(default \code{2}) Index of the coefficient to be extracted.}

\item{sort}{(default \code{TRUE}) If \code{TRUE}, results are ordered
by P-Value.}
}
\description{
Homologous methods from \code{limma}, \code{topTable} resturns a
\code{data.frame} with the \code{logFC} and  \code{PValue} per
featrue for the selcted \code{coef} and for given result (\code{rid}).
}
