% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{plotAssociation}
\alias{plotAssociation}
\alias{plotAssociation,ResultSet-methods}
\title{#' Method to add an ExposomeSet to a MultiDataSet
#'
#' This method allows to insert an object of class \link{ExposomeSet} as an
#' independent dataset into an object of class \link{MultiDataSet}.
#'
#' @name add_exp
#' @rdname add_exp-methods
#' @aliases add_exp
#' @param object An object of class \link{MultiDataSet}.
#' @param expoSet An object of class \link{ExposomeSet}.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @param ... Arguments given to \link{add_eset} from \link{MultiDataSet}.
#' @return A \link{MultiDataSet} with the \link{ExpressionSet} added as an
#' independent dataset.
#' @examples
#' data("exposome")
#' library(MultiDataSet)
#' md <- new("MultiDataSet")
#' names(md)
#' md <- add_exp(md, expo)
#' names(md)
#' @export add_exp
setGeneric("add_exp", function(object, expoSet, warnings = TRUE, ...)
    standardGeneric("add_exp")
)}
\usage{
plotAssociation(object, rid = 1, coef = 2, contrast = 1,
  type = c("manhattan", "qq"), tPV, tFC)
}
\arguments{
\item{object}{An object of class \link{ResultSet} obtained from assoc_*
functions.}

\item{rid}{(default \code{1}) Index or name of the test to be plotted. Not
used it the \code{ResultSet} comes from \link{assocSNP}.}

\item{coef}{(default \code{2}) Index of the coefficient to be extracted.}

\item{contrast}{(default \code{1}) When \code{code} corresponds to a
multicategorical variable, contasr selects the comparison.}

\item{type}{Can take \code{"manhattan"} or \code{"qq"}.}

\item{...}{Other arguments used in the different plots.}
}
\description{
#' Method to add an ExposomeClust to a MultiDataSet
#'
#' This method allows to insert an object of class \link{ExposomeClust} as an
#' independent dataset into an object of class \link{MultiDataSet}.
#'
#' @name add_cls
#' @rdname add_cls-methods
#' @aliases add_cls
#' @param object An object of class \link{MultiDataSet}.
#' @param clsSet An object of class \link{ExposomeClust}.
#' @param ... Arguments given to \link{add_eset} from \link{MultiDataSet}.
#' @return A \link{MultiDataSet} with the \link{ExpressionSet} added as an
#' independent dataset.
#' @examples
#' data("eclust")
#' library(MultiDataSet)
#' md <- new("MultiDataSet")
#' names(md)
#' md <- add_cls(md, expo_c)
#' names(md)
#' @export add_cls
setGeneric("add_cls", function(object, clsSet, ...)
    standardGeneric("add_cls")
)
}
\details{
#' Method to add an ExpressionSet to a MultiDataSet tagged as
#' protein data container.
#'
#' This method allows to insert an object of class \link{ExpressionSet} as an
#' independent dataset into an object of class \link{MultiDataSet}. The
#' \link{ExpressionSet} will be tagged, in the \link{MultiDataSet} as
#' \code{"protein"}.
#'
#' @param object An object of class \link{MultiDataSet}.
#' @param expoSet An object of class \link{ExpressionSet} with protein data.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @param ... Arguments given to \link{add_eset} from \link{MultiDataSet}.
#' @return A \link{MultiDataSet} with the \link{ExpressionSet} added as an
#' independent dataset.
#' @export add_prot
setGeneric("add_prot", function(object, expoSet, ...)
    standardGeneric("add_prot")
)
# -----------------------------------------------------------------------------
#' Method to perform an association study between transcriptome and exposom
#'
#' This function allows to perform an association study between gene
#' expression from microarray and the exposome. An \link{ExpresionSet} is
#' the object storing the gene expresion and an \link{ExposomeSet} the one
#' storing the exposome. Both of them needs to be encapsulated in a
#' \link{MultiDataSet}. The association study is perform through standard
#' \link{limma} pipeline. The function allows to perform multiple tests using
#' the argument \code{exposures}.
#'
#' @name assocGE
#' @rdname assocGE-methods
#' @aliases assocGE,MultiDataSet-rexposome
#' @param object An object of class \link{MultiDataSet} containing an
#' \code{ExposomeSet} and an \link{ExposomeSet}.
#' @param formula Model (without exposure/s) to be tested.
#' @param select One or multiple exposures existing into the
#' \link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).
#' @param set (default \code{"exposures"}) Can take values \code{"exposures"}
#' of \code{"phenotype"}.
#' @param method (default \code{"ls"}) Can takes \code{"ls"} or \code{"robust"}
#' and it will be the method used by \link{MEAL}/\code{limma} to test
#' the association.
#' @param ... Arguments passed to \code{\link{lmFit}}.
#' @param sva (default \code{FALSE}) If \code{TRUE} SVAnalysis is done.
#' @param vfilter (default \code{NULL}) Numeric number of probes used
#' in \link{sva} if argument \code{sva} is set to \code{TRUE}.de
#' @param verbose (default \code{FALSE}) If set to \code{TRUE} information
#' about the process is show.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @return An object of class \code{ResultSet} containing the results of the
#' test and the associated feature-data.
#' @export assocGE
#' @seealso \link{assocME} to test the association between epigenome and
#' exposome, \link{assocSNP} to test the association between genome and
#' exposome, \link{assocPRT} to test the association between proteome and
#' exposome, package \link{limma} for more details, \link{plotAssociation} to
#' plot the results
setGeneric("assocGE", function(object, formula, select, set = "exposures",
                               method="ls", ..., sva = FALSE, vfilter = NULL,
                               verbose = FALSE, warnings = TRUE)
    standardGeneric("assocGE")
)
#' Method to perform an association study between proteome and exposom
#'
#' This function allows to perform an association study between protein
#' expression from microarray and the exposome. An \link{ExpresionSet} is
#' the object storing the protein expresion and an \link{ExposomeSet} the one
#' storing the exposome. Both of them needs to be encapsulated in a
#' \link{MultiDataSet}. The association study is perform through standard
#' \link{limma} pipeline. The function allows to perform multiple tests using
#' the argument \code{exposures}.
#'
#' @name assocPRT
#' @rdname assocPRT-methods
#' @aliases assocPRT,MultiDataSet-rexposome
#' @param object An object of class \link{MultiDataSet} containing an
#' \code{ExposomeSet} and an \link{ExposomeSet}.
#' @param formula Model (without exposure/s) to be tested.
#' @param select One or multiple exposures existing into the
#' \link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).
#' @param set (default \code{"exposures"}) if set to \code{"phenotypes"} the
#' variables used in the association test are the phenotypes instead of the
#' exposures.
#' @param ... ...
#' @param ebayes (default \code{TRUE}) If \code{TRUE}, fitted limma modes is
#' given to \link{eBayes} before extrating results.
#' @param sva (default \code{FALSE}) if set to \code{TRUE} SVa is applied on
#' the model and the surrogate variables are incldue in the association test.
#' @param vfilter (default \code{NULL}) Numeric number of probes used
#' in \link{sva} if argument \code{sva} is set to \code{TRUE}.
#' @param ncores (default \code{1}) Number of test performed at a time.
#' @param verbose (default \code{FALSE}) If set to \code{TRUE} information
#' about the process is show.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @return An object of class \code{ResultSet} containing the results of the
#' test and the associated feature-data.
#' @export assocPRT
#' @seealso \link{assocME} to test the association between epigenome and
#' exposome, \link{assocSNP} to test the association between genome and
#' exposone, \link{assocGE} to test the association between gene expression
#' and exposome, package \link{limma} for more details,
#' \link{plotAssociation} to plot the results
setGeneric("assocPRT", function(object, formula, exposures, ..., sva = TRUE,
                                vfilter = NULL, ncores = 1,
                                verbose = FALSE, warnings = TRUE)
    standardGeneric("assocPRT")
)
#' Method to perform an association study between epigenome and exposome
#'
#' This function allows to perform epigenetic wide association studies
#' between methylation and expsome. A \link{MethylationSet} is the object
#' storing the CpGs and an \link{ExposomeSet} the one storing the exposures.
#' The association study relies on package \link{MEAL}. The function allows to
#' perform multiple tests using the argument \code{exposures}.
#'
#' @name assocME
#' @rdname assocME-methods
#' @aliases assocME,MultiDataSet-rexposome
#' @param object An object of class \link{MultiDataSet} containing an
#' \code{ExposomeSet} and a \link{MethylationSet}.
#' #' @param formula Model (without exposure/s) to be tested.
#' @param select One or multiple exposures existing into the
#' \link{ExposomeSet} (encapsulated into the \link{MultiDataSet}).
#' @param set (default \code{"exposures"}) Can take values \code{"exposures"}
#' of \code{"phenotype"}.
#' @param method (default \code{"ls"}) Can takes \code{"ls"} or \code{"robust"}
#' and it will be the method used by \link{MEAL}/\code{limma} to test
#' the association.
#' @param betas (default \code{FALSE}) By default M-values are used. If set
#' to \code{TRUE} beta-vaues are used instead.
#' @param ... Other arguments passed to \link{DAProbe} or \link{DARegion}.
#' @param sva (default \code{FALSE}) If set to \code{TRUE}, surrogate
#' variable analysis is done before calling \link{DAProbe}, or \link{DARegion},
#' and obtained surrogate variables are added.
#' @param vfilter (default \code{NULL}) Numeric number of probes used
#' in \link{sva} if argument \code{sva} is set to \code{TRUE}.
#' @param verbose (default \code{FALSE}) If set to \code{TRUE} information
#' about the process is show.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @return An object of class \code{ResultSet} containing the results of the
#' test and the associated feature-data.
#' @export assocME
#' @importClassesFrom MultiDataSet MethylationSet
#' @seealso \link{assocGE} to test the association between transcriptome
#' and exposome, \link{assocSNP} to test the association between genome and
#' exposome, package \link{MEAL} for more details, \link{plotAssociation} to
#' plot the results
setGeneric("assocME", function(object, formula, select, set="exposures",
                               method="ls", betas=FALSE,
                               ..., sva=FALSE, vfilter = NULL,
                               verbose=FALSE, warnings=TRUE)
    standardGeneric("assocME")
)
#' Method to perform an association study between genome and diseasome
#' (exposome's phenotypes)
#'
#' This function allows to perform genome wide association studies between
#' SNPs and phenotypes. A \link{SnpSet} is the object storing the genotypes and
#' an \link{ExposomeSet} the phenotypes. The association study is performed
#' through \link{snp.rhs.tests} from \link{snpStats}.
#'
#' @section Warning:
#' A difference between \code{assocSNP} and both \code{assocGE}
#' and \code{assocME}, is that \code{assocSNP} cannot perform multiple
#' test in a single run.
#'
#' @name assocSNP
#' @rdname assocSNP-methods
#' @aliases assocSNP,MultiDataSet-rexposome
#' @param object An object of class \link{MultiDataSet} containing an
#' \code{ExposomeSet} and a \link{SnpSet}.
#' @param design Formula to be tested at \link{snp.rhs.tests}.
#' @param family (default \code{"binomial"}) A string defining the generalized
#' linear model family. This can take \code{"binomial"}, \code{"Poisson"},
#' \code{"Gaussian"} or \code{"gamma"}. It is given to \link{snp.rhs.tests}.
#' @param filter.callrate (default \code{0.98}) Filter for SNPs call-rate. Ones
#' under the threshold are removed.
#' @param filter.maf (default \code{0.05}) Filter for minimum allele frequency
#' (MAF). Ones under the threshold are removed.
#' @param filter.pvalHEW (default \code{0.001}) Filter for Hardy–Weinberg. Ones
#' under the threshold are removed.
#' @param ... Argument given to \link{snp.rhs.tests} from \link{snpStats}.
#' @export assocSNP
#' @seealso \link{assocGE} to test the association between transcriptome
#' and exposome, \link{assocME} to test the association between epigenome
#' and exposome, package \link{snpStats} for more details,
#' \link{plotAssociation} to plot the results
setGeneric("assocSNP", function(object, design, family = "binomial",
                                 filter.callrate = 0.98, filter.maf = 0.05, filter.pvalHWE = 0.001, ...)
    standardGeneric("assocSNP")
)
#' Method to perform a crossomics integration analysis.
#'
#' This functions performs a crossomics analysis using all the datasets stored
#' into a \link{MultiDataSet} object. The process of integration performs a
#' sparse multiple canonical correlation analysis with all the available
#' tables. The genome (SNPs) data is transformed to a continuous variable.
#'
#' @section Warning:
#' Current accepted datasets are: \link{ExposomeSer} for exposures data.
#' \link{ExpressionSet} for micro/gene expression data, \link{MethylationSet}
#' for methylome (CpGs) data, \link{SnpSet} for genome (SNPs) data. Proteome
#' data is allowed as an \link{eSet} while tagged as \code{"proteome"} into the
#' \link{MultiDataSet} object.
#'
#' @name crossomics
#' @rdname crossomics-methods
#' @aliases crossomics,MultiDataSet-rexposome
#' @param object An object of class \link{MultiDataSet} containing the multiple
#' sets to be integrated
#' @param ncomponents (default \code{2})
#' @param na.rm (default \code{FALSE})
#' @param verbose (default \code{FALSE}) If set to \code{TRUE} information
#' about the process is show.
#' @param warnings (default \code{TRUE}) If set to \code{FALSE} warnings will
#' not be displayed.
#' @references
#' Witten, DM and Tibshirani, R and Hastie, T. A penalized matrix
#' decomposition, with applications to sparse principal components and canonical
#' correlation analysis. Biostatistics. 2009. 10.1093/biostatistics/kxp008
#'
#' Abraham, G and Inouye M. Fast Principal Component Analysis of Large-Scale
#' Genome-Wide Data. PLOS ONE. http://dx.doi.org/10.1371/journal.pone.0093766
#' @export crossomics
#' @seealso the used functions \link{MultiCCA} and \link{MultiCCA.permut} or
#' the package \link{PMA} for details, \link{plotIntegration} to plot the
#' results
setGeneric("crossomics", function(object, ncomponents = 2, ..., na.rm = FALSE,
                                  verbose = FALSE, warnings = TRUE)
    standardGeneric("crossomics")
)
Function to draw de result of an association study

This function draws two type of plots for the ResultSet from association
functions
}
\seealso{
\link{plotIntegration} for plotting integration results
}
