% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ResultSet-length.R,
%   R/ResultSet-names.R, R/ResultSet-plotAssociation.R, R/ResultSet-plotHits.R,
%   R/ResultSet-plotIntegration.R, R/ResultSet-plotLambda.R, R/ResultSet-rid.R,
%   R/ResultSet-topTable.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\alias{length,ResultSet-method}
\alias{names,ResultSet-method}
\alias{plotAssociation,ResultSet-method}
\alias{plotHits,ResultSet-method}
\alias{plotIntegration,ResultSet-method}
\alias{plotLambda,ResultSet-method}
\alias{rid,ResultSet-method}
\alias{topTable,ResultSet-method}
\title{Class ResultSet}
\usage{
\S4method{length}{ResultSet}(x)

\S4method{names}{ResultSet}(x)

\S4method{plotAssociation}{ResultSet}(object, rid = 1, coef = 2,
  contrast = 1, type = c("manhattan", "qq"), tPV, tFC)

\S4method{plotHits}{ResultSet}(object, th = 0.05, width = 0.75)

\S4method{plotIntegration}{ResultSet}(object, cmpX = 1, cmpY = 2, colors,
  lb.th = 0.2, legend.show = TRUE, ...)

\S4method{plotLambda}{ResultSet}(object, trim = 0.5, width = 0.75)

\S4method{rid}{ResultSet}(object)

\S4method{topTable}{ResultSet}(object, rid, coef = 2, contrast = 1,
  sort = TRUE)
}
\value{
An object of class \code{ResultSet}
}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \link{\code{rexposome}}. This include the
functions \code{\link{assocSNP}}, \code{\link{assocGE}},
\code{\link{assocME}}, \code{\link{assocPRT}} and \code{\link{crossomics}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Returns the amoung of analyses stored in the
\code{ResultSet}.

\item \code{names}: Returns the names of the omics data used to create
the \code{ResultSet}.

\item \code{plotAssociation}: Allows to plot a series of plots (QQ plot, Manhattan
plot and Volcano plot) depending on the results stored in the
\code{ResultSet}.

\item \code{plotHits}: Draws a bar plot with the number of hits in each
result of the \code{ResultSet}.

\item \code{plotIntegration}: Draws a plot depending on the integration analysis
stored in the \code{Resultset}.

\item \code{plotLambda}: Draws a bar plot with the lambda value of each
analyses stored in the \code{ResultSet}.

\item \code{rid}: Returns the names (aka. exposures) of the analyses
stored in the \code{ResultSet}.

\item \code{topTable}: Getter to obtain the raw \code{data.frame} from
association and integration analysis.
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{class_origin}}{Character containing the class of the objects used to
create the object.}

\item{\code{names}}{Character with the names of the datasets in the
\link{MultiDataSet} used to create the object}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{sva}}{0 if no correction process was applied. 1 if SVA was applied.
process(es).}
}}

