% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_AllClasses.R, R/ResultSet-plotHits.R,
%   R/ResultSet-plotLambda.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\alias{plotHits,ResultSet-method}
\alias{plotLambda,ResultSet-method}
\title{Class ResultSet}
\usage{
\S4method{plotHits}{ResultSet}(object, th = 0.05, width = 0.75)

\S4method{plotLambda}{ResultSet}(object, trim = 0.5, width = 0.75)
}
\value{
An object of class \code{ResultSet}
}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \link{\code{rexposome}}. This include the
functions \code{\link{assocSNP}}, \code{\link{assocGE}},
\code{\link{assocME}}, \code{\link{assocPRT}} and \code{\link{crossomics}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plotHits}: Draws a bar plot with the number of hits in each
result of the \code{ResultSet}.

\item \code{plotLambda}: Draws a bar plot with the lambda value of each
analyses stored in the \code{ResultSet}.
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{class_origin}}{Character containing the class of the objects used to
create the object.}

\item{\code{names}}{Character with the names of the datasets in the
\link{MultiDataSet} used to create the object}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{sva}}{0 if no correction process was applied. 1 if SVA was applied.
process(es).}
}}

