% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{rid}
\alias{rid}
\title{#' Function to draw de result of an integration study
#'
#' This function draws a plots for the ResultSet from integration function
#'
#' @name plotIntegration
#' @rdname plotIntegration-methods
#' @aliases plotIntegration,ResultSet-methods
#' @param object An object of class \link{ResultSet} obtained from
#' \link{crossomics}.
#' @param cmpX (default \code{1}) Value of the X-axis when ploting rsults
#' from \link{mcia}.
#' @param cmpY (default \code{2}) Value of the Y-axis when ploting rsults
#' from \link{mcia}.
#' @param tcolors (optional) Names vector with the colors sued to draw
#' each dataset. Used when ploting results from \link{MultiCCA}. If missing,
#' random colores are chosen.
#' @param lb.th (default \code{0.20}) Threshold to place labels on radar chart
#' drawn when ploting results from \link{MultiCCA}.
#' @param legend.show (default \code{TRUE}) If set to FALSE, right legend
#' of radar plot is hidden when ploting results from \link{MultiCCA}.
#' @param ... Optional arguments are given to \code{plot} from \link{omicade4}
#' pacage (argument \code{axes} is filled with values from \code{cmpX} and
#' \code{cmpY}).
#' @export plotIntegration
#' @seealso \link{plotAssociation} for plotting association results
setGeneric("plotIntegration", function(object, cmpX=1, cmpY=2, tcolors,
        lb.th=0.20, legend.show=TRUE, ...)
    standardGeneric("plotIntegration")
)
Obtain the "rid"s from a ResultSet}
\usage{
rid(object)
}
\arguments{
\item{object}{An object of class \link{ResultSet}}
}
\value{
A character vector of \code{rid}s.
}
\description{
This method resunts as character with the \code{rid}
in a given \link{ResultSet}.
}
