% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_AllMethods.R
\name{tableLambda}
\alias{tableLambda}
\title{Compute a lambda score on the results stored in a ResultSet}
\usage{
tableLambda(object, trim = 0.5)
}
\arguments{
\item{object}{An object of class \link{ResultSet}}

\item{trim}{(default \code{0.5}) percentage of right omited values for
\link{lambdaClayton}.}
}
\value{
Returns a \code{data.frame} having the exposures and the computed
lambda score.

A labeled numeric vector with the lambda score for each exposure.
}
\description{
Compute lambda score on each result in the given \link{ResultSet} by using
\code{lambdaClayton}.
}
